% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stochastic_GD.R
\name{Stochastic_GD}
\alias{Stochastic_GD}
\title{Stochastic Gradient Descent algorithm using R}
\usage{
Stochastic_GD(x, y, learning_rate)
}
\arguments{
\item{x}{the independent variable data.}

\item{y}{the dependent variable data.}

\item{learning_rate}{Rate of stochastic gradient descent}
}
\value{
Linear function
}
\description{
Fit the independent variable X and dependent variable Y linearly by  using stochastic gradient descent algorithm
}
\examples{
\dontrun{
   Gradient_Descent(c(1,2,3,4,5),c(2,4.1,5.9,7.9,10.01),0.00001)
}
}
